Added snapshot tests for some components
